backend/requirements.txt
fastapi
uvicorn
requests
pydantic
google-generativeai 
markdown2 
python-docx 
reportlab 
pypandoc 